Extract navbar to core/components
