+++ app/frontend-webapp/app/layout.tsx
@@ -7,7 +7,7 @@ import { css } from "@/styled-system/css";
 import { Metadata } from "next";
 import { Sidebar } from "@/components/core/sidebar/Sidebar";
 import { SidebarItems } from "@/config/sidebar";
-import Navbar from "@/components/Navbar";
+import Navbar from "@/components/core/Navbar";
 
 export default function RootLayout({
   children,
