+++ app/frontend-webapp/components/core/Navbar.tsx
